fix: launcher
