+++ packages/ai-chat/src/chat/components-legacy/launcher/LauncherButton.scss
@@ -1,0 +1,160 @@
+/*
+ *  Copyright IBM Corp. 2025
+ *
+ *  This source code is licensed under the Apache-2.0 license found in the
+ *  LICENSE file in the root directory of this source tree.
+ */
+
+@use "@carbon/layout";
+@use "@carbon/styles/scss/theme";
+@use "@carbon/styles/scss/motion";
+@use "@carbon/styles/scss/type";
+@use "@carbon/styles/scss/components/button" as button;
+@use "../../styles/chat-theme";
+@use "../../styles/chat-keyframes";
+
+.cds-aichat--launcher__button-container {
+  position: fixed;
+  z-index: chat-theme.$z-index;
+  border-radius: layout.$spacing-01;
+  animation: cds-aichat-launcher-in motion.$duration-moderate-01
+    motion.motion(entrance, expressive) both;
+  background-color: theme.$background;
+  box-shadow: chat-theme.$box-shadow;
+  inset-block-end: chat-theme.$launcher-position-bottom;
+  inset-inline-end: chat-theme.$launcher-position-right;
+}
+
+.cds-aichat--launcher__button-container--hidden {
+  /* Using display:none causes the entrance animation to replay */
+  visibility: hidden;
+}
+
+.cds-aichat--launcher__button-container--mobile {
+  inset-block-end: chat-theme.$launcher-position-bottom-mobile;
+  inset-inline-end: chat-theme.$launcher-position-bottom-mobile;
+}
+
+// This is the div that wraps the small launcher button. We need to add the box-shadow here because the box shadow of
+// the actual button is used for the focus state. Because the box-shadow is being added here we have to make sure to set
+// the background to transparent, otherwise there could be some weird white pixels sneaking out from beneath the
+// launcher button depending on how css decides to round the corners.
+.cds-aichat--launcher__button-container--round {
+  border-radius: 28px;
+  box-shadow: chat-theme.$box-shadow;
+}
+
+// This is the actual small launcher button itself, it's color and box shadow are already set, we just need to make it
+// round as well.
+.cds-aichat--launcher__button-container--round .cds-aichat--launcher__button {
+  border-radius: 28px;
+  background-color: chat-theme.$launcher-color-background;
+  transition: unset;
+
+  &::part(button) {
+    &:focus {
+      box-shadow: inset 0 0 0 button.$button-border-width
+        chat-theme.$launcher-color-focus-border;
+    }
+  }
+
+  &:focus {
+    border-width: button.$button-border-width;
+    border-color: chat-theme.$launcher-color-background;
+  }
+
+  &:hover {
+    border-color: chat-theme.$launcher-color-background-hover;
+    background-color: chat-theme.$launcher-color-background-hover;
+  }
+
+  &:active {
+    border-width: button.$button-border-width;
+    border-color: chat-theme.$launcher-color-background-active;
+    background-color: chat-theme.$launcher-color-background-active;
+    box-shadow: inset 0 0 0 button.$button-border-width
+      chat-theme.$launcher-color-focus-border;
+  }
+}
+
+.cds-aichat--launcher__button-container--round .cds-aichat--count-indicator {
+  position: absolute;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  padding: 0 4px;
+  border-radius: 10px;
+  background-color: chat-theme.$unread-indicator-color-background;
+  color: chat-theme.$unread-indicator-color-text;
+  @include type.type-style("caption-01");
+
+  inset-block-start: calc(-1 * #{layout.$spacing-01});
+  inset-inline-end: calc(-1 * #{layout.$spacing-01});
+  min-block-size: 20px;
+
+  // This is 20px and not a multiple of 4 because of the 12px font-size plus 4px of padding on each side.
+  min-inline-size: 20px;
+}
+
+.cds-aichat--container--render[dir="rtl"] .cds-aichat--count-indicator {
+  inset-inline: calc(-1 * #{layout.$spacing-01}) unset;
+}
+
+.cds-aichat--launcher__button-container--no-animation {
+  animation: none;
+}
+
+.cds-aichat--launcher__button-container--bounce-animation {
+  animation: cds-aichat-launcher-bounce 500ms
+    motion.motion(entrance, expressive) forwards;
+}
+
+cds-button.cds-aichat--launcher__button {
+  &::part(button) {
+    svg {
+      block-size: 24px;
+      fill: chat-theme.$launcher-color-avatar;
+      inline-size: 24px;
+    }
+
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    padding: inherit;
+    border-radius: inherit;
+    block-size: inherit;
+    inline-size: inherit;
+    max-inline-size: inherit;
+  }
+
+  &::part(button):focus {
+    border-width: button.$button-border-width;
+  }
+
+  position: static;
+  padding: 0;
+  border-radius: layout.$spacing-01;
+  background-color: chat-theme.$launcher-color-background;
+  block-size: chat-theme.$launcher-default-size;
+  inline-size: chat-theme.$launcher-default-size;
+  transition:
+    background 250ms ease-in-out,
+    transform 150ms ease;
+}
+
+.cds-aichat--launcher__svg {
+  fill: currentcolor;
+}
+
+.cds-aichat--launcher__avatar {
+  border-radius: 50%;
+  block-size: 32px;
+  inline-size: 32px;
+  user-select: none;
+}
+
+.cds-aichat--launcher__button .cds-aichat__count-indicator {
+  box-shadow: 1px layout.$spacing-01 layout.$spacing-01 rgb(23 23 23 / 30%);
+  inset-block-start: calc(-1 * #{layout.$spacing-02});
+  inset-inline-end: calc(-1 * #{layout.$spacing-02});
+}
