+++ packages/ai-chat/src/chat/components-legacy/launcher/LauncherMobileContainer.scss
@@ -1,0 +1,180 @@
+/*
+ *  Copyright IBM Corp. 2025
+ *
+ *  This source code is licensed under the Apache-2.0 license found in the
+ *  LICENSE file in the root directory of this source tree.
+ */
+
+@use "@carbon/layout";
+@use "@carbon/styles/scss/theme";
+@use "@carbon/styles/scss/motion";
+@use "@carbon/styles/scss/type";
+@use "@carbon/styles/scss/components/button" as button;
+@use "../../styles/chat-theme";
+@use "../../styles/chat-keyframes";
+@use "../../styles/chat-config" as chat-config;
+
+.cds-aichat--launcher-extended__container {
+  block-size: chat-theme.$launcher-default-size;
+  inline-size: chat-theme.$launcher-default-size;
+}
+
+cds-button.cds-aichat--launcher-extended__button::part(button) {
+  padding: 0;
+  border-width: button.$button-border-width;
+  block-size: 100%;
+  cursor: pointer;
+  inline-size: 100%;
+}
+
+.cds-aichat--launcher-extended__button--hidden {
+  visibility: hidden;
+}
+
+.cds-aichat--launcher-extended__wrapper-container {
+  position: relative;
+  overflow: hidden;
+  block-size: 100%;
+  inline-size: 100%;
+}
+
+.cds-aichat--launcher-extended__wrapper {
+  position: absolute;
+  display: flex;
+  inline-size: calc(#{chat-theme.$launcher-extended-width} - 6px);
+  inset-inline-end: 0;
+  max-inline-size: calc(100vw - chat-theme.$launcher-position-bottom - 6px);
+}
+
+.cds-aichat--launcher-extended__text-holder {
+  position: relative;
+  flex: 1;
+  padding: layout.$spacing-04 0 layout.$spacing-04 layout.$spacing-04;
+  text-align: start;
+}
+
+.cds-aichat--launcher-extended__greeting {
+  position: absolute;
+  display: flex;
+  align-items: center;
+  block-size: 100%;
+  color: chat-theme.$launcher-mobile-color-text;
+  @include type.type-style("body-01");
+
+  inline-size: calc(100% - 12px);
+  inset-block-end: 0;
+  word-break: break-word;
+}
+
+.cds-aichat-is-phone .cds-aichat--launcher-extended__greeting {
+  @include type.type-style("body-01");
+}
+
+.cds-aichat--launcher-extended__greeting-text {
+  display: -webkit-box;
+  overflow: hidden;
+  -webkit-box-orient: vertical;
+  -webkit-line-clamp: 2;
+  line-clamp: 2;
+  text-overflow: ellipsis;
+}
+
+.cds-aichat--launcher-extended__button .cds-aichat--launcher__icon-holder {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  block-size: calc(
+    chat-theme.$launcher-default-size - button.$button-border-width * 2
+  );
+  inline-size: calc(
+    chat-theme.$launcher-default-size - button.$button-border-width * 2
+  );
+  margin-inline-start: auto;
+
+  svg {
+    position: absolute;
+    block-size: 24px;
+    inline-size: 24px;
+  }
+}
+
+.cds-aichat--launcher-extended__button
+  .cds-aichat--launcher-extended__element--hidden {
+  display: none;
+}
+
+.cds-aichat--launcher-extended__button
+  .cds-aichat--launcher-extended__element--fade-out {
+  animation: #{chat-config.$prefix}-launcher-extended-element-fade-out 500ms
+    ease-out 400ms both;
+}
+
+.cds-aichat--launcher-extended__button
+  .cds-aichat--launcher-extended__element--fade-in {
+  animation: #{chat-config.$prefix}-launcher-extended-element-fade-in 500ms
+    ease-out 400ms both;
+}
+
+.cds-aichat--launcher__button-container.cds-aichat--launcher-extended__button--extended {
+  border-radius: 14px;
+  inline-size: chat-theme.$launcher-extended-width;
+
+  cds-button.cds-aichat--launcher__button {
+    inline-size: chat-theme.$launcher-extended-width;
+  }
+
+  cds-button.cds-aichat--launcher__button.cds-aichat--launcher-extended__button::part(
+      button
+    ) {
+    border-radius: 14px;
+  }
+
+  .cds-aichat--launcher__svg {
+    inline-size: 24px;
+  }
+}
+
+.cds-aichat--launcher__button-container.cds-aichat--launcher-extended__button--extended-animation {
+  animation:
+    #{chat-config.$prefix}-launcher-extend 300ms
+      cubic-bezier(0.25, 0.46, 0.45, 0.94) 500ms both,
+    #{chat-config.$prefix}-launcher-partially-round 300ms ease-in 500ms both;
+
+  cds-button.cds-aichat--launcher-extended__button::part(button) {
+    animation: #{chat-config.$prefix}-launcher-partially-round 300ms ease-in
+      500ms both;
+  }
+
+  .cds-aichat--launcher__svg {
+    animation: #{chat-config.$prefix}-launcher-icon-to-24 300ms ease-out 500ms
+      both;
+  }
+}
+
+.cds-aichat--launcher__button-container.cds-aichat--launcher-extended__button--reduced-animation {
+  animation:
+    #{chat-config.$prefix}-launcher-reduce 300ms
+      cubic-bezier(0.25, 0.46, 0.45, 0.94) 500ms both,
+    #{chat-config.$prefix}-launcher-completely-round 300ms ease-out 500ms both;
+
+  cds-button.cds-aichat--launcher-extended__button::part(button) {
+    animation: #{chat-config.$prefix}-launcher-completely-round 300ms ease-out
+      500ms both;
+  }
+
+  .cds-aichat--launcher__svg {
+    animation: #{chat-config.$prefix}-launcher-icon-to-32 300ms ease-out 500ms
+      both;
+  }
+}
+
+.cds-aichat--container--render[dir="rtl"]
+  .cds-aichat--launcher-extended__container {
+  .cds-aichat--launcher-extended__wrapper {
+    inset-inline: 0 unset;
+  }
+
+  .cds-aichat--launcher-extended__text-holder {
+    padding: layout.$spacing-04 layout.$spacing-04 layout.$spacing-04 0;
+  }
+}
