+++ packages/ai-chat/src/chat/components-legacy/launcher/LauncherMobileContainer.tsx
@@ -7,16 +7,22 @@
  *  @license
  */
 
+import CDSButton from "@carbon/web-components/es/components/button/button.js";
+import ChatLaunch24 from "@carbon/icons/es/chat--launch/24.js";
+import { carbonIconToReact } from "../../utils/carbonIcon";
 import cx from "classnames";
 import React, {
-  RefObject,
+  MutableRefObject,
   useCallback,
   useEffect,
   useRef,
   useState,
 } from "react";
+import { useIntl } from "react-intl";
 import { useSelector } from "../../hooks/useSelector";
 
+import { useAriaAnnouncer } from "../../hooks/useAriaAnnouncer";
+import { useLanguagePack } from "../../hooks/useLanguagePack";
 import { useOnMount } from "../../hooks/useOnMount";
 import { usePrevious } from "../../hooks/usePrevious";
 import { useServiceManager } from "../../hooks/useServiceManager";
@@ -24,10 +30,13 @@ import actions from "../../store/actions";
 import { AppState } from "../../../types/state/AppState";
 import { BOUNCING_ANIMATION_TIMEOUTS } from "../../../types/config/LauncherConfig";
 import { setAnimationTimeouts } from "../../utils/animationUtils";
+import { LauncherButton, LauncherHandle } from "./LauncherButton";
 import {
-  LauncherExtended,
-  LauncherExtendedFunctions,
-} from "./LauncherExtended";
+  calculateAndSetMaxExtendedLauncherWidth,
+  checkIfUserSwipedRight,
+  doFadeAnimationForElements,
+} from "./launcherUtils";
+import type { LauncherTouchStartCoordinates } from "./launcherUtils";
 
 interface LauncherMobileContainerProps {
   onToggleOpen: () => void;
@@ -35,7 +44,7 @@ interface LauncherMobileContainerProps {
   /**
    * Necessary to get access to the ref created within App.tsx.
    */
-  launcherRef: RefObject<LauncherExtendedFunctions>;
+  launcherRef: MutableRefObject<LauncherHandle | null>;
 
   /**
    * If the main Carbon AI Chat window is open the launcher should be hidden.
@@ -46,7 +55,15 @@ interface LauncherMobileContainerProps {
 function LauncherMobileContainer(props: LauncherMobileContainerProps) {
   const { launcherRef, onToggleOpen, launcherHidden } = props;
   const serviceManager = useServiceManager();
+  const languagePack = useLanguagePack();
+  const intl = useIntl();
+  const ariaAnnouncer = useAriaAnnouncer();
   const { launcher } = useSelector((state: AppState) => state.config.derived);
+  const launcherAvatarURL = useSelector((state: AppState) =>
+    state.config.derived.themeWithDefaults.aiEnabled
+      ? undefined
+      : state.config.derived.launcher.mobile.avatarUrlOverride,
+  );
   const unreadHumanAgentCount = useSelector(
     (state: AppState) => state.humanAgentState.numUnreadMessages,
   );
@@ -61,6 +78,8 @@ function LauncherMobileContainer(props: LauncherMobileContainerProps) {
 
   const [isStartingBounceAnimation, setIsStartingBounceAnimation] =
     useState(false);
+  const [animateExtendedState, setAnimateExtendedState] = useState(false);
+  const [showGreetingMessage, setShowGreetingMessage] = useState(false);
   const prevIsExtended = usePrevious(isExtended);
   const prevWasReduced = usePrevious(wasReduced);
   // The bounce turn start off on in the recurring animation flow. We only care about the initial value and not its
@@ -68,17 +87,25 @@ function LauncherMobileContainer(props: LauncherMobileContainerProps) {
   // the flow.
   const initialBounceTurn = useRef(bounceTurn).current;
   const previouslyPlayedExtendAnimation = useRef(wasReduced).current;
-  const extendLauncherTimeoutIDRef = useRef(null);
-  const reduceLauncherTimeoutIDRef = useRef(null);
-  const endBounceAnimationRef = useRef(null);
+  const extendLauncherTimeoutIDRef = useRef<NodeJS.Timeout>();
+  const endBounceAnimationRef = useRef<() => void>();
   const shouldBounceRef = useRef(
     previouslyPlayedExtendAnimation && !disableBounce,
   );
 
+  const textHolderRef = useRef<HTMLDivElement>();
+  const greetingMessageRef = useRef<HTMLDivElement>();
+  const extendedContainerRef = useRef<HTMLDivElement>();
+  const buttonRef = useRef<CDSButton>();
+  const touchStartRef = useRef<LauncherTouchStartCoordinates>({
+    touchStartX: null,
+    touchStartY: null,
+  });
+
   const { timeToExpand } = launcher.mobile;
   const isExpandedLauncherEnabled = launcher.mobile.isOn;
 
-  // If the launcher container mounted with the mobile launcher not in the extended state, and it's previous value is
+  // If the launcher container mounted with the mobile launcher not in the extended state, and its previous value is
   // undefined, this means the launcher should be in the extended state playing the extended animation if not in the
   // tooling preview.
   const playExtendAnimation = prevIsExtended === undefined && !isExtended;
@@ -92,13 +119,36 @@ function LauncherMobileContainer(props: LauncherMobileContainerProps) {
   const disableIntroAnimation =
     isExtending || hasReduced || isStartingBounceAnimation;
 
+  const shouldReduceExtendedLauncher = !isExtended && prevIsExtended;
+  const extendWithAnimation = isExtended && animateExtendedState;
+  const extendWithoutAnimation = isExtended && !animateExtendedState;
+  const launcherGreetingMessage =
+    launcher.mobile.title || languagePack.launcher_mobileGreeting;
+
+  const ChatLaunch = carbonIconToReact(ChatLaunch24);
+  const launcherAvatar = launcherAvatarURL ? (
+    // eslint-disable-next-line jsx-a11y/alt-text
+    <img
+      className="cds-aichat--launcher__avatar"
+      src={launcherAvatarURL}
+      aria-hidden
+      alt=""
+    />
+  ) : (
+    <ChatLaunch className="cds-aichat--launcher__svg" />
+  );
+
+  useEffect(() => {
+    setAnimateExtendedState(playExtendAnimation);
+  }, [playExtendAnimation]);
+
   const setLauncherStateAsReduced = useCallback(() => {
     if (!wasReduced) {
       serviceManager.store.dispatch(
         actions.setLauncherProperty("mobileLauncherWasReduced", true),
       );
     }
-  }, [wasReduced, serviceManager]);
+  }, [serviceManager.store, wasReduced]);
 
   // This function kicks off the process of reducing the extended launcher, such as when the user scrolls the page, by
   // setting mobileLauncherIsExtended in launcher state to false. If the user does scroll the page, it will be tracked.
@@ -112,7 +162,7 @@ function LauncherMobileContainer(props: LauncherMobileContainerProps) {
         actions.setLauncherProperty("mobileLauncherIsExtended", false),
       );
     }
-  }, [isExtended, serviceManager]);
+  }, [isExtended, serviceManager.store]);
 
   const setExpandAnimationTimeout = useCallback(() => {
     // Begin timeout to set launcher in the extended state.
@@ -135,7 +185,7 @@ function LauncherMobileContainer(props: LauncherMobileContainerProps) {
     const endBounceAnimation = endBounceAnimationRef.current;
     if (endBounceAnimation) {
       endBounceAnimation();
-      endBounceAnimationRef.current = null;
+      endBounceAnimationRef.current = undefined;
     }
 
     // Prevent the launcher from bouncing if it was toggled and allowed to play the bounce animation.
@@ -149,20 +199,21 @@ function LauncherMobileContainer(props: LauncherMobileContainerProps) {
     setLauncherStateAsReduced();
   }, [reduceLauncher, serviceManager.store, setLauncherStateAsReduced]);
 
-  // When the launcher mounts, we should determine if it should prepare to play the "extend" animation, or kickoff the
-  // bounce animation. We should kick off the bounce animation early if we have to so that we can easily determine
-  // later on if it should be canceled.
+  // When the launcher mounts, we should determine if it should prepare to play the "extend" animation, or kick off the
+  // bounce animation. We should kick off the bounce animation early if we have to so that we can easily determine later
+  // on if it should be canceled.
   useOnMount(() => {
     // Determine if the mobile launcher wasn't reduced and can play the "extend" animation.
     if (!wasReduced && playExtendAnimation && isExpandedLauncherEnabled) {
       setExpandAnimationTimeout();
     } else if (shouldBounceRef.current) {
       const launcherContainerElement =
-        launcherRef?.current?.launcherContainerElement();
+        launcherRef.current?.launcherContainerElement?.() ??
+        extendedContainerRef.current;
 
       if (launcherContainerElement) {
         const startRecurringBounceAnimation = () => {
-          // This function is added as an event listener to the container, however the function isn't actually run until
+          // This function is added as an event listener to the container; however, the function isn't actually run until
           // the event listener is triggered. Because of this it's possible that the state has since changed, and we
           // actually don't want to bounce the launcher after all, so we need to check that we still want to bounce.
           if (shouldBounceRef.current) {
@@ -211,59 +262,188 @@ function LauncherMobileContainer(props: LauncherMobileContainerProps) {
   });
 
   // If the main window has been opened then clear all timers and set the launcher state as if it had been
-  // clicked open. This is to protect against scenarios where the main window is  opened using other methods
-  // besides clicking on the launcher.
+  // clicked open. This is to protect against scenarios where the main window is opened using other methods besides
+  // clicking on the launcher.
   useEffect(() => {
     if (viewState.mainWindow) {
       // Clear timers and update launcher state so that no more greeting messages or bounces occur.
       setDefaultLauncherState();
     }
-  }, [viewState, setDefaultLauncherState]);
+  }, [setDefaultLauncherState, viewState]);
 
   function clearTimeouts() {
     const extendLauncherTimeoutID = extendLauncherTimeoutIDRef.current;
-    const reduceLauncherTimeoutID = reduceLauncherTimeoutIDRef.current;
 
-    // Clears all the existing timeouts that were set.
-    if (reduceLauncherTimeoutID) {
-      clearTimeout(reduceLauncherTimeoutID);
-      reduceLauncherTimeoutIDRef.current = null;
-    }
     if (extendLauncherTimeoutID) {
       clearTimeout(extendLauncherTimeoutID);
-      extendLauncherTimeoutIDRef.current = null;
+      extendLauncherTimeoutIDRef.current = undefined;
     }
   }
 
-  // When the launcher is toggled, reduce the launcher if it's extended and open the widget.
   const handleToggleOpen = useCallback(() => {
     setDefaultLauncherState();
     onToggleOpen();
   }, [onToggleOpen, setDefaultLauncherState]);
 
-  // Track the user swiping to the right over the launcher and reduce the launcher.
-  const handleSwipeRight = useCallback(() => {
-    reduceLauncher();
-  }, [reduceLauncher]);
+  const containerClassName = cx("cds-aichat--launcher-extended__container", {
+    "cds-aichat--launcher__button-container--mobile": true,
+    "cds-aichat--launcher-extended__button--extended": extendWithoutAnimation,
+    "cds-aichat--launcher-extended__button--extended-animation":
+      extendWithAnimation,
+    "cds-aichat--launcher-extended__button--reduced-animation":
+      shouldReduceExtendedLauncher,
+    "cds-aichat--launcher__button-container--no-animation":
+      disableIntroAnimation,
+  });
+
+  useEffect(() => {
+    const textHolderElement = textHolderRef.current;
+    const greetingMessageElement = greetingMessageRef.current;
+    const extendedContainerElement = extendedContainerRef.current;
+
+    if (
+      !textHolderElement ||
+      !greetingMessageElement ||
+      !extendedContainerElement
+    ) {
+      return;
+    }
+
+    calculateAndSetMaxExtendedLauncherWidth(
+      textHolderElement,
+      greetingMessageElement,
+      extendedContainerElement,
+    );
+  }, [ariaAnnouncer, extendWithoutAnimation, launcherGreetingMessage]);
+
+  useEffect(() => {
+    const buttonElement = buttonRef.current;
+    const containerElement = extendedContainerRef.current;
+
+    if (isExtended && buttonElement) {
+      if (animateExtendedState) {
+        doFadeAnimationForElements(
+          { fadeInElement: greetingMessageRef.current, fadeInTime: 300 },
+          () => {
+            setAnimateExtendedState(false);
+          },
+        );
+      } else {
+        setShowGreetingMessage(true);
+      }
+
+      document.addEventListener("scroll", reduceLauncher);
+
+      const handleTouchMove = (event: TouchEvent) => {
+        checkIfUserSwipedRight(
+          event.touches[0],
+          touchStartRef.current,
+          reduceLauncher,
+        );
+      };
+
+      const handleTouchStart = (event: TouchEvent) => {
+        const { clientX, clientY } = event.touches[0];
+        const touchStart = touchStartRef.current;
+
+        touchStart.touchStartX = clientX;
+        touchStart.touchStartY = clientY;
+
+        buttonRef.current?.addEventListener("touchmove", handleTouchMove);
+      };
+
+      buttonElement.addEventListener("touchstart", handleTouchStart);
+
+      return () => {
+        buttonElement.removeEventListener("touchmove", handleTouchMove);
+        buttonElement.removeEventListener("touchstart", handleTouchStart);
+        document.removeEventListener("scroll", reduceLauncher);
+      };
+    }
+
+    if (shouldReduceExtendedLauncher && containerElement) {
+      const reduceAnimationEndListener = () => {
+        setLauncherStateAsReduced();
+        setAnimateExtendedState(true);
+        containerElement.removeEventListener(
+          "animationend",
+          reduceAnimationEndListener,
+        );
+      };
+
+      containerElement.addEventListener(
+        "animationend",
+        reduceAnimationEndListener,
+      );
+
+      doFadeAnimationForElements({
+        fadeOutElement: greetingMessageRef.current,
+      });
+
+      return () => {
+        containerElement.removeEventListener(
+          "animationend",
+          reduceAnimationEndListener,
+        );
+      };
+    }
+
+    return undefined;
+  }, [
+    animateExtendedState,
+    isExtended,
+    reduceLauncher,
+    setLauncherStateAsReduced,
+    shouldReduceExtendedLauncher,
+  ]);
+
+  if (launcherRef.current) {
+    launcherRef.current.launcherContainerElement = () =>
+      extendedContainerRef.current ?? undefined;
+  }
 
   return (
-    <LauncherExtended
-      className={cx({
-        "cds-aichat--launcher__button-container--no-animation":
-          disableIntroAnimation,
-      })}
+    <LauncherButton
+      className="cds-aichat--launcher-extended__button"
+      buttonRef={buttonRef}
+      containerClassName={containerClassName}
+      containerRef={extendedContainerRef}
+      closedLabel={languagePack.launcher_isClosed}
+      intl={intl}
+      openLabel={languagePack.launcher_isOpen}
+      launcherHidden={launcherHidden}
+      onToggleOpen={handleToggleOpen}
       ref={launcherRef}
-      launcher={launcher}
-      isMobile={true}
       showUnreadIndicator={showUnreadIndicator}
       unreadHumanAgentCount={unreadHumanAgentCount}
-      isExtended={isExtended}
-      playExtendAnimation={playExtendAnimation}
-      onToggleOpen={handleToggleOpen}
-      onSwipeRight={handleSwipeRight}
-      onReduceEnd={setLauncherStateAsReduced}
-      launcherHidden={launcherHidden}
-    />
+    >
+      <div className="cds-aichat--launcher-extended__wrapper-container">
+        <div className="cds-aichat--launcher-extended__wrapper">
+          <div
+            className="cds-aichat--launcher-extended__text-holder"
+            ref={textHolderRef}
+          >
+            <div
+              className={cx("cds-aichat--launcher-extended__greeting", {
+                "cds-aichat--launcher-extended__element--hidden":
+                  !showGreetingMessage,
+              })}
+              ref={greetingMessageRef}
+            >
+              <div
+                className="cds-aichat--launcher-extended__greeting-text"
+                aria-hidden={!isExtended}
+              >
+                {launcherGreetingMessage}
+              </div>
+            </div>
+          </div>
+          <div className="cds-aichat--launcher__icon-holder">
+            {launcherAvatar}
+          </div>
+        </div>
+      </div>
+    </LauncherButton>
   );
 }
