+++ packages/ai-chat/src/chat/components-legacy/launcher/_imports.scss
@@ -5,4 +5,6 @@
  *  LICENSE file in the root directory of this source tree.
  */
 
-@use "Launcher";
+@use "LauncherButton";
+@use "LauncherDesktopContainer";
+@use "LauncherMobileContainer";
